shibboleth destroy session and redirect to shibboleth logout

+++ commsy/pages/context_logout.php
@@ -35,8 +35,17 @@ $flash = $session->getValue('flash');
 if ( $session->issetValue('root_session_id') ) {
    $root_session_id = $session->getValue('root_session_id');
 }
-$session_manager->delete($SID,true);
-$session->reset();
+if ($environment->getConfiguration('c_shibboleth_direct_login') and !empty($environment->getConfiguration('c_shibboleth_redirect_url'))){
+	if ($_SERVER['Shib_userId'] != $session->getValue('user_id')){
+		$session_manager->delete($SID,true);
+		$session->reset();
+		redirect_with_url($environment->getConfiguration('c_shibboleth_redirect_url'));
+	}
+} else {
+	$session_manager->delete($SID,true);
+	$session->reset();
+}
+
 
 include_once('classes/cs_session_item.php');
 $session = new cs_session_item();
